+++ src/Grpc/Interop/test/testassets/InteropWebsite/Program.cs
@@ -17,6 +17,7 @@
 #endregion
 
 using Microsoft.AspNetCore.Server.Kestrel.Core;
+using Microsoft.Extensions.Logging.Console;
 
 namespace InteropTestsWebsite;
 
@@ -31,11 +32,9 @@ public class Program
         Host.CreateDefaultBuilder(args)
             .ConfigureLogging(builder =>
             {
-                builder.AddConsole(loggerOptions =>
+                builder.AddSimpleConsole(loggerOptions =>
                 {
-#pragma warning disable CS0618 // Type or member is obsolete
-                    loggerOptions.DisableColors = true;
-#pragma warning restore CS0618 // Type or member is obsolete
+                    loggerOptions.ColorBehavior = LoggerColorBehavior.Disabled;
                 });
                 builder.SetMinimumLevel(LogLevel.Trace);
             })
